Allow for setting minimum distance before flipsnap starts
I wanted to control the distance per Flipsnap implementation before it starts as it was 'too eager' on its default hard-coded setting of 5. I use flipsnap in conjunction with iScroll so when vertically scrolling an element a user can still move on the x axis and if done too much (i.e. > 5) results in a confusing x-axis change when they meant to only scroll on the y-axis. I've put the value up to 50 for my own needs but this can easily be changed.